requirements.txt
requests
pandas
matplotlib
numpy
selenium
mysql-connector-python
